Add files via upload
The snap shot is the the output of the project. The image is of different situation when the Driver is sleeping and when the Driver is active. when the Driver is active then score is increasing and when he/she closes the eyes the score is decreasing. if it closed for longer period then alarm will be generated.